feat: 🎸 分析しづらくてスターを消した
